Hand back a callable no-op where the consumer-owned lane has no nil
A non-optional closure return degraded to a zeroed pair, which Swift traps on
when it wraps the result, so a collected implementation killed the process on
the lane whose whole point is that it does not. Non-optional Swift pointers had
the opposite problem — their zero is a well-formed bit pattern that is not a
value of the type — so they now take the unsatisfiable terminal instead, and the
three sibling fan-out receivers that still fell through to a bare return take the
same two-lane terminal as their no-sibling siblings.

+++ src/Swift.Runtime/src/Swift/Runtime/InteropServices/SwiftMarshal.cs
@@ -1527,11 +1527,16 @@ public static class SwiftMarshal
             {
                 unsafe
                 {
-                    int caseValue = swiftSize switch
+                    // Read unsigned. A discriminator is an unsigned index into the case list, so a
+                    // signed read of the top half of a width (case 0x8000 of a 2-byte enum, say)
+                    // sign-extends to a negative value and names a case that does not exist. This
+                    // matches ReadDiscriminator, which the tuple/container element path uses for the
+                    // same bytes.
+                    ulong caseValue = swiftSize switch
                     {
                         1 => ((byte*)swiftSource)[0],
-                        2 => *(short*)swiftSource,
-                        _ => *(int*)swiftSource,
+                        2 => *(ushort*)swiftSource,
+                        _ => *(uint*)swiftSource,
                     };
                     return (T)Enum.ToObject(typeof(T), caseValue);
                 }
